added time select feature

+++ covidTracker.py
@@ -4,30 +4,29 @@ import requests
 from pymongo import MongoClient
 import pprint
 from bson.son import SON
+from datetime import date, timedelta
 
 def main():
-    # covidDataURL = 'https://covidtracking.com/api/v1/states/daily.json'
-    # statesDataURL = 'https://raw.githubusercontent.com/nytimes/covid-19-data/master/us-counties.csv'
-    # credsFile = 'credentials.json'
+    covidDataURL = 'https://covidtracking.com/api/v1/states/daily.json'
+    statesDataURL = 'https://raw.githubusercontent.com/nytimes/covid-19-data/master/us-counties.csv'
+    credsFile = 'credentials.json'
     configFile = 'trackerConfig.json'
-    # db = getDB(credsFile)
-    # updateDB(db,covidDataURL, statesDataURL)
     config = configure(configFile)
+    db = getDB(credsFile)
+    refresh(config['refresh'], db,covidDataURL, statesDataURL)
     pipeline = generate_pipeline(config)
     print(pipeline)
-    # covid = db['covid']
-    # pprint.pprint(list(db.covid.aggregate(pipeline)))
-
-
+    if config['collection'] == 'states':
+        pprint.pprint(list(db.states.aggregate(pipeline)))
+    else:
+        pprint.pprint(list(db.covid.aggregate(pipeline)))
+        
 #this is real the real work is. Takes in the config file and creates a pipeline based on the contents of said file.
 def generate_pipeline(config):
     pipeline = []
-    pipeline = [interpret_counties(config)]
+    pipeline = [interpret_counties(config) , interpret_time(config)]
     #example pipeline to make sure everything is working
-    #pipeline = [{"$match": {"state": "CA"}},
-    #            {"$match": {"date": {"$gte": 20200401, "$lte": 20200415}}},
-    #            {"$project": {"_id":0, "positive":1, "date":1}},
-    #            {"$sort": {"date":1}}]
+    #pipeline = [{"$match": {"state": "CA"}},{"$match": {"date": {"$gte": 20200401, "$lte": 20200415}}},{"$project": {"_id":0, "positive":1, "date":1}},{"$sort": {"date":1}}]
     return pipeline
 
 
@@ -36,8 +35,28 @@ def generate_pipeline(config):
 def interpret_collection(collection):
     pass
 
-def interpret_time(time):
-    pass
+def interpret_time(config):
+    time = config['time']
+    today = date.today()
+    pipe = ""
+    if time == "today":
+        pipe = {"$match":{"date": today.strftime('%Y%m%d')}}
+    elif time == "yesterday":
+        yesterday = today - timedelta(days=1)
+        pipe = {"$match":{"date":yesterday.strftime('%Y%m%d')}}
+    elif time == "week":
+        week_ago = today - timedelta(weeks=1)
+        pipe = {"$match":{"date":{"$gte":week_ago.strftime('%Y%m%d'),  "$lte":today.strftime('%Y%m%d')}}}
+    elif time == "month":
+        pipe = {"$match":{"date":{"$gte":today.strftime('%Y%m01'), "$lte":today.strftime('%Y%m%d')}}}
+    else:
+        start = time['start']
+        end = time['end']
+        pipe = {"$match":{"date":{"$gte":start , "$lte":end}}}
+    return pipe  
+
+
+    
 
 def interpret_target(target):
     pass
@@ -59,15 +78,18 @@ def interpret_analysis(analysis):
 
 
 #updates db collections with data from APIs
-def updateDB(db, covidDataURL, statesDataURL):
-    covidStr = requests.get(covidDataURL).text
-    statesCSV = requests.get(statesDataURL).text
-    covid = json.loads(covidStr)
-    states = csv2json(statesCSV)
-    db.covid.drop()
-    db.states.drop()
-    db.covid.insert_many(covid)
-    db.states.insert_many(states)
+def refresh(refresh_bool, db, covidDataURL, statesDataURL):
+    if refresh_bool:
+        covidStr = requests.get(covidDataURL).text
+        statesCSV = requests.get(statesDataURL).text
+        covid = json.loads(covidStr)
+        states = csv2json(statesCSV)
+        db.covid.drop()
+        db.states.drop()
+        db.covid.insert_many(covid)
+        db.states.insert_many(states)
+    else:
+        pass
 
 def csv2json(csv):
     newLis = []
@@ -76,7 +98,10 @@ def csv2json(csv):
         line = line.split(",")
         dic = {}
         for item in range(len(header)):
-            dic[header[item]] = line[item]
+            if item == 0:
+                dic[header[item]] = "".join(line[0].split("-"))
+            else:
+                dic[header[item]] = line[item]
         newLis.append(dic)  
     return(newLis)
